// src/lib/generated/types/JobInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JobState } from "./JobState";
import type { Policy } from "./Policy";

export type JobInfo = { id: string, title: string, state: JobState, policy: Policy, done: number, total: number, label: string | null, error: string | null, };